feat(StringCalculator): 커스텀 구분자로 입력된 값의 합을 구하는 기능 구현
