+++ README.md
@@ -13,7 +13,7 @@
 1) 구분자를 가진 문자열 입력 받는다.
 
 2) 구분자를 기준으로 분리 한다.
-  - 커스텀 구분자(문자열 앞부분의 “//”와 “\n” 사이에 위치하는 문자)로 분리를 시도한다.
+  -X 커스텀 구분자(문자열 앞부분의 “//”와 “\n” 사이에 위치하는 문자)로 분리를 시도한다.
   -X 커스텀 구분자가 없으면  기본 구분자(,|:)로 분리한다.
 
 3) 구분자를 기준으로 분리한 숫자 입력 값을 검증 한다.
